Improve log message and add git diff functionality
In 'gitgen.rs', additional log messages were added to print out the changes and the git diff. Also, a new function 'get_git_diff' was introduced to 'ops.rs' to fetch the difference between the working directory and the latest commit. This will provide better debugging and understanding of code changes.