Add 'async-trait' dependency and implement ChatClient trait
The 'async-trait' dependency was introduced to the project, which is used to implement the 'ChatClient' trait on the 'GPTClient' struct in 'msg.rs'.  Also, enhancements on `std::path::Path` in `gitgen.rs` were made for `.env` file checking. This trait abstracts common methods related to sending messages, making it easier to create new chat clients in the future.